Use `ClassValue` for storing field data
Seems to be better suited for this use case and might even allow class
unloading when the associated value refers to the class, which is
currently the case for `FieldData` (which contains the `Field`, which
contains a reference to the declaring class).
Alternatively could consider refactoring `FieldData` to not store the
corresponding `Field` anymore but instead only its relevant information.